KAN-47 fix: fix input styles, features move to src/features folder
