+++ src/features/auth/forms/CreateNewPassword/index.ts
